refactor: refactor dayjs formatDate helper
