+++ src/shared/lib/utils/formatters/dateTime/date/index.ts
@@ -1,0 +1,1 @@
+export { DateFormats, getDate } from './getDate';
