+++ src/shared/lib/utils/formatters/dateTime/index.ts
@@ -1,0 +1,1 @@
+export { useFormatDateTime } from './useFormatDateTime';
