+++ src/shared/lib/utils/formatters/dateTime/useFormatDateTime.ts
@@ -1,0 +1,7 @@
+import { getDate } from './date';
+
+export function useFormatDateTime() {
+  return {
+    getDate,
+  };
+}
