--- src/shared/lib/utils/formatters/duration/formatDuration.ts
@@ -1,5 +1,0 @@
-import * as dateTime from './time';
-
-export function useFormatDuration() {
-  return dateTime;
-}
