+++ src/shared/lib/utils/formatters/duration/time/index.ts
@@ -1,4 +1,3 @@
-export { DateFormats, getDate } from './getDate';
 export { getDays } from './getDays';
 export { getHours } from './getHours';
 export { getMinutes } from './getMinutes';
