+++ src/shared/lib/utils/formatters/duration/useFormatDuration.ts
@@ -1,0 +1,10 @@
+import { getDays, getHours, getMinutes, getSeconds } from './time';
+
+export function useFormatDuration() {
+  return {
+    getDays,
+    getHours,
+    getMinutes,
+    getSeconds,
+  };
+}
